Fixed Some YouTrack Issues

+++ src/main/java/fun/rubicon/util/SafeMessage.java
@@ -9,7 +9,6 @@ package fun.rubicon.util;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.Message;
 import net.dv8tion.jda.core.entities.MessageEmbed;
-import net.dv8tion.jda.core.entities.PrivateChannel;
 import net.dv8tion.jda.core.entities.TextChannel;
 
 import java.util.concurrent.TimeUnit;
@@ -30,7 +29,6 @@ public class SafeMessage {
     }
 
 
-
     public static Message sendMessageBlocking(TextChannel textChannel, Message message) {
         if (hasPermissions(textChannel))
             return textChannel.sendMessage(message).complete();
@@ -38,12 +36,12 @@ public class SafeMessage {
     }
 
     public static void sendMessage(TextChannel textChannel, String message) {
-        if (hasPermissions(textChannel))
+        if (hasPermissions(textChannel) && message.toCharArray().length < 2000)
             textChannel.sendMessage(message).queue();
     }
 
     public static void sendMessage(TextChannel textChannel, String message, int deleteTime) {
-        if (hasPermissions(textChannel) && hasDeletePermission(textChannel))
+        if (hasPermissions(textChannel) && hasDeletePermission(textChannel) && message.toCharArray().length < 2000)
             textChannel.sendMessage(message).queue(msg -> msg.delete().queueAfter(deleteTime, TimeUnit.SECONDS));
         else if (hasPermissions(textChannel))
             textChannel.sendMessage(message).queue();
@@ -63,10 +61,11 @@ public class SafeMessage {
     }
 
     public static Message sendMessageBlocking(TextChannel textChannel, String message) {
-        if (hasPermissions(textChannel))
+        if (hasPermissions(textChannel) && message.toCharArray().length < 2000)
             return textChannel.sendMessage(message).complete();
         return null;
     }
+
     public static Message sendMessageBlocking(TextChannel textChannel, MessageEmbed message) {
         if (hasPermissions(textChannel))
             return textChannel.sendMessage(message).complete();
@@ -77,7 +76,6 @@ public class SafeMessage {
     private static boolean hasPermissions(TextChannel channel) {
         return channel.getGuild().getSelfMember().hasPermission(channel, Permission.MESSAGE_READ) && channel.getGuild().getSelfMember().hasPermission(channel, Permission.MESSAGE_WRITE);
     }
-
 
 
     private static boolean hasDeletePermission(TextChannel channel) {
